Add and display the badge list
I created a component that will be used to show the stored badges on the page. This list will show the avatar, the full name of the attendant, Twitter, and its job title. For the display of this component, I decided to create a page that shows the conference logo, the created list, and a button to create a badge for a new attendant.